+++ README.md
@@ -2,10 +2,12 @@
 Yet another GRID attempt, for the WPI Hackathon this time.
 
 # TODO
+Reorganize server file structure
 Fix game file recognition and loading
 Rework client connection code
 Create a text-based client
 Provide just-in-time game creation
 Game lifecycle - ready up, set up, play!
 Game replays and archive?
-Fix chess specifically
+Fix chess specifically (later? / in-depth task)
+Server announcements API for the homepage
