+++ clients/WS_Client/public/logo.png
