+++ games/Chess/pieces/bishop_black.png
