+++ games/Chess/pieces/bishop_white.png
