+++ games/Chess/pieces/knight_black.png
