+++ games/Chess/pieces/pawn_black.png
