+++ games/Chess/pieces/pawn_white.png
